# Conflicts:
#	gieFrontend/src/main/java/module-info.java
#	gieFrontend/src/main/java/org/giefront/TestFront.java
#	gieFrontend/src/main/resources/org/giefront/fournisseur.fxml
#	gieFrontend/src/main/resources/org/giefront/fournisseurEntro.fxml

+++ gieBackend/src/main/java/org/gieback/DAO/EntrepriseDao.java
@@ -85,7 +85,7 @@ public class EntrepriseDao implements IEntrepriseDao {
     }
 
     @Override
-    public void deleteByid(int id) {
+    public void deleteByid(Long id) {
         String hql = "delete from Entreprise where id =:id";
         try {
             entityManager.getTransaction().begin();
